+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# unrealeased
+
+Breaking
+  * AZ-460: Rewrite module for TF 0.13+
+
 # v1.0.0 - 2019-06-12
 
 Added